+++ src/collective/cookiecuttr/upgrades.py
@@ -7,9 +7,9 @@ from zope.component import queryUtility
 
 import logging
 
-PROFILE_ID = 'profile-collective.cookiecuttr:default'
-UNINSTALL = 'profile-collective.cookiecuttr:uninstall'
-DEPENDENCY = 'profile-collective.z3cform.datagridfield:default'
+PROFILE_ID = "profile-collective.cookiecuttr:default"
+UNINSTALL = "profile-collective.cookiecuttr:uninstall"
+DEPENDENCY = "profile-collective.z3cform.datagridfield:default"
 
 
 def upgrade_to_0002(context, logger=None):
@@ -18,57 +18,74 @@ def upgrade_to_0002(context, logger=None):
 
     if logger is None:
         # Called as upgrade step: define our own logger.
-        logger = logging.getLogger('collective.cookiecuttr')
+        logger = logging.getLogger("collective.cookiecuttr")
 
     # Get the existing values
     registry = queryUtility(IRegistry)
     text = registry.get(
-        'collective.cookiecuttr.interfaces.ICookieCuttrSettings.text',
-        u' '
+        "collective.cookiecuttr.interfaces.ICookieCuttrSettings.text", u" "
     )
     link = registry.get(
-        'collective.cookiecuttr.interfaces.ICookieCuttrSettings.link',
-        u' '
+        "collective.cookiecuttr.interfaces.ICookieCuttrSettings.link", u" "
     )
     accept = registry.get(
-        'collective.cookiecuttr.interfaces.ICookieCuttrSettings.accept_button',
-        u' '
+        "collective.cookiecuttr.interfaces.ICookieCuttrSettings.accept_button", u" "
     )
 
     if not text:
-        text = u' '
+        text = u" "
 
     if not link:
-        link = u' '
+        link = u" "
 
     if not accept:
-        accept = u' '
+        accept = u" "
 
     # re-import profile
-    portal_setup = getToolByName(context, 'portal_setup')
-    portal_setup.runImportStepFromProfile(PROFILE_ID, 'plone.app.registry')
+    portal_setup = getToolByName(context, "portal_setup")
+    portal_setup.runImportStepFromProfile(PROFILE_ID, "plone.app.registry")
 
     # save preexisting data in a dict
-    lt = getToolByName(context, 'portal_languages')
+    lt = getToolByName(context, "portal_languages")
     lang = safe_unicode(lt.getDefaultLanguage())
-    registry['collective.cookiecuttr.interfaces.ICookieCuttrSettings.text'] = [dict(language=lang, text=text)]
-    registry['collective.cookiecuttr.interfaces.ICookieCuttrSettings.link'] = [dict(language=lang, text=link)]
-    registry['collective.cookiecuttr.interfaces.ICookieCuttrSettings.accept_button'] = [dict(language=lang, text=accept)]
+    registry["collective.cookiecuttr.interfaces.ICookieCuttrSettings.text"] = [
+        dict(language=lang, text=text)
+    ]
+    registry["collective.cookiecuttr.interfaces.ICookieCuttrSettings.link"] = [
+        dict(language=lang, text=link)
+    ]
+    registry["collective.cookiecuttr.interfaces.ICookieCuttrSettings.accept_button"] = [
+        dict(language=lang, text=accept)
+    ]
 
     # install datagridfield
     portal_setup.runAllImportStepsFromProfile(DEPENDENCY)
 
-    logger.info('Done')
+    logger.info("Done")
 
 
 def upgrade_from_0002_to_0003(context, logger=None):
     """Add option to move cookie message to the bottom to the control panel."""
 
     if logger is None:
-        logger = logging.getLogger('collective.cookiecuttr')
+        logger = logging.getLogger("collective.cookiecuttr")
 
     # Re-import plone.app.registry
-    portal_setup = getToolByName(context, 'portal_setup')
-    portal_setup.runImportStepFromProfile(PROFILE_ID, 'plone.app.registry')
+    portal_setup = getToolByName(context, "portal_setup")
+    portal_setup.runImportStepFromProfile(PROFILE_ID, "plone.app.registry")
 
-    logger.info('Successfully upgraded from 0002 to 0003')
+    logger.info("Successfully upgraded from 0002 to 0003")
+
+
+def upgrade_from_0003_to_0004(context, logger=None):
+    """Add option to move cookie message to the bottom to the control panel."""
+
+    if logger is None:
+        logger = logging.getLogger("collective.cookiecuttr")
+
+    # Re-import rolemap to allow Site Administrators acess to the control panel
+    portal_setup = getToolByName(context, "portal_setup")
+    portal_setup.runImportStepFromProfile(PROFILE_ID, "rolemap")
+    portal_setup.runImportStepFromProfile(PROFILE_ID, "controlpanel")
+
+    logger.info("Successfully upgraded from 0003 to 0004")