+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Also allow Site Administrators access to the Control Panel, not just Managers.
+  [fredvd]
 
 
 1.0.1 (2020-01-17)